Edit the admin profile in a dialog
The profile is a short form reached from the user menu, so a full settings
page was more navigation than the task warranted — opening it lost the
merchant's place in the dashboard. It now opens as a dialog over the current
page and the /settings/profile route is removed.

The form lives in the app package because its hooks, schema and language list
do, so the top bar takes a handler for the menu item rather than owning the
dialog itself.
