+++ packages/dashboard-starter/src/routeTree.gen.ts
@@ -34,7 +34,6 @@ import { Route as SettingsStoreRouteImport } from './../../dashboard/src/routes/
 import { Route as SettingsStockLocationsRouteImport } from './../../dashboard/src/routes/_authenticated/$storeId/settings/stock-locations'
 import { Route as SettingsStaffRouteImport } from './../../dashboard/src/routes/_authenticated/$storeId/settings/staff'
 import { Route as SettingsReasonsRouteImport } from './../../dashboard/src/routes/_authenticated/$storeId/settings/reasons'
-import { Route as SettingsProfileRouteImport } from './../../dashboard/src/routes/_authenticated/$storeId/settings/profile'
 import { Route as SettingsProductTypesRouteImport } from './../../dashboard/src/routes/_authenticated/$storeId/settings/product-types'
 import { Route as SettingsPaymentMethodsRouteImport } from './../../dashboard/src/routes/_authenticated/$storeId/settings/payment-methods'
 import { Route as SettingsMarketsRouteImport } from './../../dashboard/src/routes/_authenticated/$storeId/settings/markets'
@@ -194,11 +193,6 @@ const SettingsStaffRoute = SettingsStaffRouteImport.update({
 const SettingsReasonsRoute = SettingsReasonsRouteImport.update({
   id: '/reasons',
   path: '/reasons',
-  getParentRoute: () => SettingsRoute,
-} as any)
-const SettingsProfileRoute = SettingsProfileRouteImport.update({
-  id: '/profile',
-  path: '/profile',
   getParentRoute: () => SettingsRoute,
 } as any)
 const SettingsProductTypesRoute = SettingsProductTypesRouteImport.update({
@@ -420,7 +414,6 @@ export interface FileRoutesByFullPath {
   '/$storeId/settings/markets': typeof SettingsMarketsRoute
   '/$storeId/settings/payment-methods': typeof SettingsPaymentMethodsRoute
   '/$storeId/settings/product-types': typeof SettingsProductTypesRoute
-  '/$storeId/settings/profile': typeof SettingsProfileRoute
   '/$storeId/settings/reasons': typeof SettingsReasonsRoute
   '/$storeId/settings/staff': typeof SettingsStaffRoute
   '/$storeId/settings/stock-locations': typeof SettingsStockLocationsRoute
@@ -480,7 +473,6 @@ export interface FileRoutesByTo {
   '/$storeId/settings/markets': typeof SettingsMarketsRoute
   '/$storeId/settings/payment-methods': typeof SettingsPaymentMethodsRoute
   '/$storeId/settings/product-types': typeof SettingsProductTypesRoute
-  '/$storeId/settings/profile': typeof SettingsProfileRoute
   '/$storeId/settings/reasons': typeof SettingsReasonsRoute
   '/$storeId/settings/staff': typeof SettingsStaffRoute
   '/$storeId/settings/stock-locations': typeof SettingsStockLocationsRoute
@@ -544,7 +536,6 @@ export interface FileRoutesById {
   '/_authenticated/$storeId/settings/markets': typeof SettingsMarketsRoute
   '/_authenticated/$storeId/settings/payment-methods': typeof SettingsPaymentMethodsRoute
   '/_authenticated/$storeId/settings/product-types': typeof SettingsProductTypesRoute
-  '/_authenticated/$storeId/settings/profile': typeof SettingsProfileRoute
   '/_authenticated/$storeId/settings/reasons': typeof SettingsReasonsRoute
   '/_authenticated/$storeId/settings/staff': typeof SettingsStaffRoute
   '/_authenticated/$storeId/settings/stock-locations': typeof SettingsStockLocationsRoute
@@ -608,7 +599,6 @@ export interface FileRouteTypes {
     | '/$storeId/settings/markets'
     | '/$storeId/settings/payment-methods'
     | '/$storeId/settings/product-types'
-    | '/$storeId/settings/profile'
     | '/$storeId/settings/reasons'
     | '/$storeId/settings/staff'
     | '/$storeId/settings/stock-locations'
@@ -668,7 +658,6 @@ export interface FileRouteTypes {
     | '/$storeId/settings/markets'
     | '/$storeId/settings/payment-methods'
     | '/$storeId/settings/product-types'
-    | '/$storeId/settings/profile'
     | '/$storeId/settings/reasons'
     | '/$storeId/settings/staff'
     | '/$storeId/settings/stock-locations'
@@ -731,7 +720,6 @@ export interface FileRouteTypes {
     | '/_authenticated/$storeId/settings/markets'
     | '/_authenticated/$storeId/settings/payment-methods'
     | '/_authenticated/$storeId/settings/product-types'
-    | '/_authenticated/$storeId/settings/profile'
     | '/_authenticated/$storeId/settings/reasons'
     | '/_authenticated/$storeId/settings/staff'
     | '/_authenticated/$storeId/settings/stock-locations'
@@ -940,13 +928,6 @@ declare module '@tanstack/react-router' {
       path: '/reasons'
       fullPath: '/$storeId/settings/reasons'
       preLoaderRoute: typeof SettingsReasonsRouteImport
-      parentRoute: typeof SettingsRoute
-    }
-    '/_authenticated/$storeId/settings/profile': {
-      id: '/_authenticated/$storeId/settings/profile'
-      path: '/profile'
-      fullPath: '/$storeId/settings/profile'
-      preLoaderRoute: typeof SettingsProfileRouteImport
       parentRoute: typeof SettingsRoute
     }
     '/_authenticated/$storeId/settings/product-types': {
@@ -1208,7 +1189,6 @@ interface SettingsRouteChildren {
   SettingsMarketsRoute: typeof SettingsMarketsRoute
   SettingsPaymentMethodsRoute: typeof SettingsPaymentMethodsRoute
   SettingsProductTypesRoute: typeof SettingsProductTypesRoute
-  SettingsProfileRoute: typeof SettingsProfileRoute
   SettingsReasonsRoute: typeof SettingsReasonsRoute
   SettingsStaffRoute: typeof SettingsStaffRoute
   SettingsStockLocationsRoute: typeof SettingsStockLocationsRoute
@@ -1231,7 +1211,6 @@ const SettingsRouteChildren: SettingsRouteChildren = {
   SettingsMarketsRoute: SettingsMarketsRoute,
   SettingsPaymentMethodsRoute: SettingsPaymentMethodsRoute,
   SettingsProductTypesRoute: SettingsProductTypesRoute,
-  SettingsProfileRoute: SettingsProfileRoute,
   SettingsReasonsRoute: SettingsReasonsRoute,
   SettingsStaffRoute: SettingsStaffRoute,
   SettingsStockLocationsRoute: SettingsStockLocationsRoute,
