+++ packages/dashboard/src/routeTree.gen.ts
@@ -33,7 +33,6 @@ import { Route as SettingsStoreRouteImport } from './routes/_authenticated/$stor
 import { Route as SettingsStockLocationsRouteImport } from './routes/_authenticated/$storeId/settings/stock-locations'
 import { Route as SettingsStaffRouteImport } from './routes/_authenticated/$storeId/settings/staff'
 import { Route as SettingsReasonsRouteImport } from './routes/_authenticated/$storeId/settings/reasons'
-import { Route as SettingsProfileRouteImport } from './routes/_authenticated/$storeId/settings/profile'
 import { Route as SettingsProductTypesRouteImport } from './routes/_authenticated/$storeId/settings/product-types'
 import { Route as SettingsPaymentMethodsRouteImport } from './routes/_authenticated/$storeId/settings/payment-methods'
 import { Route as SettingsMarketsRouteImport } from './routes/_authenticated/$storeId/settings/markets'
@@ -187,11 +186,6 @@ const SettingsStaffRoute = SettingsStaffRouteImport.update({
 const SettingsReasonsRoute = SettingsReasonsRouteImport.update({
   id: '/reasons',
   path: '/reasons',
-  getParentRoute: () => SettingsRoute,
-} as any)
-const SettingsProfileRoute = SettingsProfileRouteImport.update({
-  id: '/profile',
-  path: '/profile',
   getParentRoute: () => SettingsRoute,
 } as any)
 const SettingsProductTypesRoute = SettingsProductTypesRouteImport.update({
@@ -407,7 +401,6 @@ export interface FileRoutesByFullPath {
   '/$storeId/settings/markets': typeof SettingsMarketsRoute
   '/$storeId/settings/payment-methods': typeof SettingsPaymentMethodsRoute
   '/$storeId/settings/product-types': typeof SettingsProductTypesRoute
-  '/$storeId/settings/profile': typeof SettingsProfileRoute
   '/$storeId/settings/reasons': typeof SettingsReasonsRoute
   '/$storeId/settings/staff': typeof SettingsStaffRoute
   '/$storeId/settings/stock-locations': typeof SettingsStockLocationsRoute
@@ -465,7 +458,6 @@ export interface FileRoutesByTo {
   '/$storeId/settings/markets': typeof SettingsMarketsRoute
   '/$storeId/settings/payment-methods': typeof SettingsPaymentMethodsRoute
   '/$storeId/settings/product-types': typeof SettingsProductTypesRoute
-  '/$storeId/settings/profile': typeof SettingsProfileRoute
   '/$storeId/settings/reasons': typeof SettingsReasonsRoute
   '/$storeId/settings/staff': typeof SettingsStaffRoute
   '/$storeId/settings/stock-locations': typeof SettingsStockLocationsRoute
@@ -527,7 +519,6 @@ export interface FileRoutesById {
   '/_authenticated/$storeId/settings/markets': typeof SettingsMarketsRoute
   '/_authenticated/$storeId/settings/payment-methods': typeof SettingsPaymentMethodsRoute
   '/_authenticated/$storeId/settings/product-types': typeof SettingsProductTypesRoute
-  '/_authenticated/$storeId/settings/profile': typeof SettingsProfileRoute
   '/_authenticated/$storeId/settings/reasons': typeof SettingsReasonsRoute
   '/_authenticated/$storeId/settings/staff': typeof SettingsStaffRoute
   '/_authenticated/$storeId/settings/stock-locations': typeof SettingsStockLocationsRoute
@@ -589,7 +580,6 @@ export interface FileRouteTypes {
     | '/$storeId/settings/markets'
     | '/$storeId/settings/payment-methods'
     | '/$storeId/settings/product-types'
-    | '/$storeId/settings/profile'
     | '/$storeId/settings/reasons'
     | '/$storeId/settings/staff'
     | '/$storeId/settings/stock-locations'
@@ -647,7 +637,6 @@ export interface FileRouteTypes {
     | '/$storeId/settings/markets'
     | '/$storeId/settings/payment-methods'
     | '/$storeId/settings/product-types'
-    | '/$storeId/settings/profile'
     | '/$storeId/settings/reasons'
     | '/$storeId/settings/staff'
     | '/$storeId/settings/stock-locations'
@@ -708,7 +697,6 @@ export interface FileRouteTypes {
     | '/_authenticated/$storeId/settings/markets'
     | '/_authenticated/$storeId/settings/payment-methods'
     | '/_authenticated/$storeId/settings/product-types'
-    | '/_authenticated/$storeId/settings/profile'
     | '/_authenticated/$storeId/settings/reasons'
     | '/_authenticated/$storeId/settings/staff'
     | '/_authenticated/$storeId/settings/stock-locations'
@@ -909,13 +897,6 @@ declare module '@tanstack/react-router' {
       path: '/reasons'
       fullPath: '/$storeId/settings/reasons'
       preLoaderRoute: typeof SettingsReasonsRouteImport
-      parentRoute: typeof SettingsRoute
-    }
-    '/_authenticated/$storeId/settings/profile': {
-      id: '/_authenticated/$storeId/settings/profile'
-      path: '/profile'
-      fullPath: '/$storeId/settings/profile'
-      preLoaderRoute: typeof SettingsProfileRouteImport
       parentRoute: typeof SettingsRoute
     }
     '/_authenticated/$storeId/settings/product-types': {
@@ -1170,7 +1151,6 @@ interface SettingsRouteChildren {
   SettingsMarketsRoute: typeof SettingsMarketsRoute
   SettingsPaymentMethodsRoute: typeof SettingsPaymentMethodsRoute
   SettingsProductTypesRoute: typeof SettingsProductTypesRoute
-  SettingsProfileRoute: typeof SettingsProfileRoute
   SettingsReasonsRoute: typeof SettingsReasonsRoute
   SettingsStaffRoute: typeof SettingsStaffRoute
   SettingsStockLocationsRoute: typeof SettingsStockLocationsRoute
@@ -1193,7 +1173,6 @@ const SettingsRouteChildren: SettingsRouteChildren = {
   SettingsMarketsRoute: SettingsMarketsRoute,
   SettingsPaymentMethodsRoute: SettingsPaymentMethodsRoute,
   SettingsProductTypesRoute: SettingsProductTypesRoute,
-  SettingsProfileRoute: SettingsProfileRoute,
   SettingsReasonsRoute: SettingsReasonsRoute,
   SettingsStaffRoute: SettingsStaffRoute,
   SettingsStockLocationsRoute: SettingsStockLocationsRoute,
